sidebar finished
