+++ src/App.jsx
@@ -1,5 +1,7 @@
 import { BrowserRouter as Router } from "react-router-dom";
 import Routes from "./router/routes";
+import "./App.css";
+
 function App() {
   return (
     <Router>
