+++ src/pages/index.js
@@ -1,0 +1,5 @@
+import Profile from "./profile/profile";
+import TeamPage from "./team/team";
+import FeedbackPage from "./feedback/feedback";
+
+export { Profile, TeamPage, FeedbackPage };
