+++ src/router/path.jsx
@@ -1,0 +1,5 @@
+export const PATH = {
+  Profile: "/profile",
+  TeamPage: "/team",
+  Feedback: "/feedback",
+};
